Fixes on the csproj files and renamed ForkTests
