+++ src/Sierra.sln
@@ -47,12 +47,14 @@ Global
 		{17909DD2-D384-4E32-9904-1387B726B2D2}.Release|x64.ActiveCfg = Release|x64
 		{17909DD2-D384-4E32-9904-1387B726B2D2}.Release|x64.Build.0 = Release|x64
 		{17909DD2-D384-4E32-9904-1387B726B2D2}.Release|x64.Deploy.0 = Release|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|x64.ActiveCfg = Debug|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|x64.Build.0 = Debug|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|Any CPU.ActiveCfg = Release|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|x64.ActiveCfg = Release|x64
-		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|x64.Build.0 = Release|x64
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Debug|x64.Build.0 = Debug|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|x64.ActiveCfg = Release|Any CPU
+		{D14121D9-5924-4AD0-9F41-795DE446B1F8}.Release|x64.Build.0 = Release|Any CPU
 		{1CE8C239-1376-4A8F-8D14-8218BDC24CFB}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{1CE8C239-1376-4A8F-8D14-8218BDC24CFB}.Debug|x64.ActiveCfg = Debug|x64
 		{1CE8C239-1376-4A8F-8D14-8218BDC24CFB}.Debug|x64.Build.0 = Debug|x64
