+++ src/Tests/Sierra.Actor.Tests/Sierra.Actor.Tests.csproj
@@ -1,11 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
     <Platform>x64</Platform>
     <Platforms>x64</Platforms>
     <IsPackable>false</IsPackable>
-   
   </PropertyGroup>
 
   <ItemGroup>
